Restructure DRA e2e tests into baseline/extended layout
