+++ hack/testing/e2e-common.sh
@@ -710,7 +710,7 @@ function cluster_kueue_deploy {
             deploy_with_certmanager "$1"
         fi
     elif [[ -n ${DRA_EXAMPLE_DRIVER_VERSION:-} ]]; then
-        build_and_apply_kueue_manifests "$1" "${ROOT_DIR}/test/e2e/config/dra"
+        build_and_apply_kueue_manifests "$1" "${ROOT_DIR}/test/e2e/config/dra/baseline"
     elif [ "$E2E_USE_HELM" == 'true' ]; then
         helm_install "$1" "${ROOT_DIR}/test/e2e/config/default/values.yaml"
     elif [[ ${E2E_TARGET_FOLDER:-} == "multikueue/sequential" ]]; then
