+++ test/e2e/config/dra/baseline/manager_e2e_patch.yaml
