+++ test/e2e/dra/baseline/dra_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package dra
+package baseline
 
 import (
 	"github.com/onsi/ginkgo/v2"
